requirements.txt
#boto3~=1.24.65
#botocore~=1.27.65
Flask~=2.2.2
Flask_Bcrypt~=1.0.1
Flask_Cors~=3.0.10
Flask_SQLAlchemy~=2.5.1
minio~=7.1.11
pause~=0.3
persist-queue~=0.8.0
PyJWT~=2.6.0
pythonnet~=2.5.2
requests~=2.28.1
s3path~=0.3.4
waitress~=2.1.2
watchdog~=2.1.9
tomli >= 1.1.0 ; python_version < "3.11"